+++ docs/CHANGELOG.rst
@@ -6,6 +6,15 @@ All notable changes to this project are documented in this file.
 This project adheres to `Semantic Versioning <http://semver.org/>`_.
 
 
+==========
+Unreleased
+==========
+
+Fix
+---
+- Fix slow annotation value query by forcing inner query evaluation
+
+
 ===================
 41.0.0 - 2024-11-18
 ===================